feat(plugin): modularize bridge.rs into sub modules
